同步一些老改动 视觉通信协议修改 | 增加正弦扫频时直接给电机输入电流的逻辑 | 修改正弦扫频函数

+++ modules/algorithm/crc8.h
@@ -5,7 +5,7 @@
 
 #define CRC_START_8 0x00
 
-uint8_t crc_8(const uint8_t *input_str, uint16_t num_bytes);
+uint8_t crc_8(const uint8_t *input_str, size_t num_bytes);
 uint8_t update_crc_8(uint8_t crc, uint8_t val);
 
 #endif
